moved symbol to sym.rs
